renamed existing module to pbtx-key-provider; added base app in order to have build and release capability for the module
