+++ build.gradle
@@ -1,149 +1,11 @@
-apply plugin: 'com.android.library'
-
-apply plugin: 'kotlin-android'
-
-apply plugin: 'kotlin-android-extensions'
-
-apply plugin: 'com.google.protobuf'
-
-
-//Remove 'com.jfrog.artifactory' plugin if you are not using Artifactory
-//apply plugin: 'com.jfrog.artifactory'
-
-apply plugin: 'maven-publish'
-
-// apply plugin: 'org.jetbrains.dokka-android'
-sourceSets {
-    main.java.srcDirs += "${protobuf.generatedFilesBaseDir}/main/javalite"
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
+plugins {
+    id 'com.android.application' version '7.1.3' apply false
+    id 'com.android.library' version '7.1.3' apply false
+    id 'org.jetbrains.kotlin.android' version '1.5.30' apply false
+    id 'com.google.protobuf' version '0.8.18' apply false
 }
 
-
-android {
-    compileSdkVersion 29
-    defaultConfig {
-        minSdkVersion 26
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    // Needed to get bitcoin-j to produce a valid apk for android.
-    packagingOptions {
-        exclude 'lib/x86_64/darwin/libscrypt.dylib'
-        exclude 'lib/x86_64/freebsd/libscrypt.so'
-        exclude 'lib/x86_64/linux/libscrypt.so'
-    }
-
-    lintOptions {
-        disable 'InvalidPackage'
-    }
-
-
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
-
-repositories {
-    google()
-    jcenter()
-}
-
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.31"
-    implementation 'androidx.core:core-ktx:1.0.2'
-
-    implementation 'one.block:eosiojava:1.0.0'
-
-    testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'androidx.test:runner:1.2.0'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-
-    implementation 'io.grpc:grpc-protobuf:1.25.0'
-    implementation 'com.googlecode.playn:playn-webgl:1.3.1'
-
-    implementation 'com.google.crypto.tink:tink-android:1.6.1'
-
-
-}
-
-def libraryGroupId = 'ekis'
-def libraryArtifactId = 'ekisandroidkeystoresignatureprovider'
-def libraryVersion = '1.0.0'
-
-task androidSourcesJar(type: Jar) {
-    classifier = 'sources'
-    from android.sourceSets.main.java.srcDirs
-}
-
-
-protobuf {
-    protoc {
-        artifact = 'com.google.protobuf:protoc:3.8.0-rc-1'
-    }
-    generateProtoTasks {
-        all().each { task ->
-            task.builtins {
-                java {
-                    option "lite"
-                }
-            }
-        }
-    }
-}
-
-
-project.afterEvaluate {
-    publishing {
-        publications {
-            maven(MavenPublication) {
-                groupId libraryGroupId
-                version libraryVersion
-
-
-                pom.withXml {
-                    final dependenciesNode = asNode().appendNode('dependencies')
-
-                    ext.addDependency = { dep, scope ->
-                        if (dep.group == null || dep.version == null || dep.name == null || dep.name == "unspecified")
-                            return // ignore invalid dependencies
-
-                        final dependencyNode = dependenciesNode.appendNode('dependency')
-                        dependencyNode.appendNode('groupId', dep.group)
-                        dependencyNode.appendNode('artifactId', dep.name)
-                        dependencyNode.appendNode('version', dep.version)
-                        dependencyNode.appendNode('scope', scope)
-
-                        if (!dep.transitive) {
-                            // If this dependency is transitive, we should force exclude all its dependencies them from the POM
-                            final exclusionNode = dependencyNode.appendNode('exclusions').appendNode('exclusion')
-                            exclusionNode.appendNode('groupId', '*')
-                            exclusionNode.appendNode('artifactId', '*')
-                        } else if (!dep.properties.excludeRules.empty) {
-                            // Otherwise add specified exclude rules
-                            final exclusionNode = dependencyNode.appendNode('exclusions').appendNode('exclusion')
-                            dep.properties.excludeRules.each { rule ->
-                                exclusionNode.appendNode('groupId', rule.group ?: '*')
-                                exclusionNode.appendNode('artifactId', rule.module ?: '*')
-                            }
-                        }
-                    }
-
-                    // List all "compile" dependencies (for old Gradle)
-                    configurations.compile.getDependencies().each { dep -> addDependency(dep, "compile") }
-                    // List all "api" dependencies (for new Gradle) as "compile" dependencies
-                    configurations.api.getDependencies().each { dep -> addDependency(dep, "compile") }
-                    // List all "implementation" dependencies (for new Gradle) as "runtime" dependencies
-                    configurations.implementation.getDependencies().each { dep -> addDependency(dep, "runtime") }
-                }
-            }
-        }
-    }
-}
-
