+++ pbtx-key-provider/src/main/AndroidManifest.xml
