+++ pbtx-key-provider/src/main/java/com/pbtx/Model/KeyModel.java
@@ -1,0 +1,25 @@
+package com.pbtx.Model;
+
+import com.pbtx.PublicKey;
+
+public class KeyModel {
+
+    PublicKey key;
+    String alias;
+
+    public PublicKey getKey() {
+        return key;
+    }
+
+    public void setKey(PublicKey key) {
+        this.key = key;
+    }
+
+    public String getAlias() {
+        return alias;
+    }
+
+    public void setAlias(String alias) {
+        this.alias = alias;
+    }
+}
